support "extended nodes"
This converts some variables to their unsidged types, which enables
the port to load all maps which use the regular Doom BSP nodes format.

This enables the port to load some maps that it would normally choke
upon, e.g. Nova Maps 30 and 31; Eviternity Map 15 and Sunlust Maps 16,
18, 20, 24, 28 and 31.

It does still not support DeePBSP and ZDBSP nodes, and will probably
never do.

Fixes #23.

+++ Source/r_defs.h
@@ -181,10 +181,10 @@ typedef struct line_s
 {
   vertex_t *v1, *v2;     // Vertices, from v1 to v2.
   fixed_t dx, dy;        // Precalculated v2 - v1 for side checking.
-  short flags;           // Animation related.
+  unsigned short flags;           // Animation related.
   short special;         
   short tag;
-  short sidenum[2];      // Visual appearance: SideDefs.
+  unsigned short sidenum[2];      // Visual appearance: SideDefs.
   fixed_t bbox[4];       // A bounding box, for the linedef's extent
   slopetype_t slopetype; // To aid move clipping.
   sector_t *frontsector; // Front and back sector.
@@ -206,7 +206,7 @@ typedef struct line_s
 typedef struct subsector_s
 {
   sector_t *sector;
-  short numlines, firstline;
+  int numlines, firstline;
 } subsector_t;
 
 // phares 3/14/98
@@ -262,7 +262,7 @@ typedef struct
 {
   fixed_t  x,  y, dx, dy;        // Partition line.
   fixed_t bbox[2][4];            // Bounding box for each child.
-  unsigned short children[2];    // If NF_SUBSECTOR its a subsector.
+  int children[2];    // If NF_SUBSECTOR its a subsector.
 } node_t;
 
 // posts are runs of non masked source pixels
